+++ api/app/main.py
@@ -128,6 +128,7 @@ from .routes_admin_privacy import router as admin_privacy_router
 from .routes_admin_qrpack import router as admin_qrpack_router
 from .routes_admin_qrposter_pack import router as admin_qrposter_router
 from .routes_admin_support import router as admin_support_router
+from .routes_admin_flags import router as admin_flags_router
 from .routes_admin_webhooks import router as admin_webhooks_router
 from .routes_alerts import router as alerts_router
 from .routes_analytics_outlets import router as analytics_outlets_router
@@ -1005,6 +1006,7 @@ app.include_router(troubleshoot_router)
 app.include_router(help_router)
 app.include_router(support_router)
 app.include_router(admin_support_router)
+app.include_router(admin_flags_router)
 app.include_router(staff_support_router)
 app.include_router(support_console_router)
 app.include_router(admin_webhooks_router)